+++ src/Application/Async/Http.php
@@ -204,7 +204,7 @@ final class Http implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
-    public function notFoundRequestHandler(callable $handle): self
+    public function routeNotFound(callable $handle): self
     {
         return new self(
             $this->os,